+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - Administrator is extended
     - Transport is extended
     - Product is extended
+    - Brand is extended
 
 ### [shopsys/shopsys]
 #### Added